% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_risk_index.R
\name{home_risk_index}
\alias{home_risk_index}
\title{Index to assess risk of home loss in the event of a wildfire}
\usage{
home_risk_index(index_data)
}
\arguments{
\item{defensible_space}{is the area around a home or other structure that has been modified to reduce fire hazard. Options are between 0 to +100ft.}

\item{fire_proofing_degree}{amount of fire proofing done by a household to prevent things like embers from entering the home or reduce flamability of the house.
values ranked from 0 (none) to 5 (very high).}

\item{temp}{temperature (degrees F)}

\item{humidity}{humidity (percent)}

\item{wind_speed}{wind speed (mph)}
}
\value{
Ranking (Very Low to Extreme) of ow vulnerable a home is being damaged by wildfire during a wildfire event.
}
\description{
Scientists predict more frequent and more severe wildifre events in the event of climate chage. However, many areas
are increasing in population size which results in developers and cities building into the Wildland Urban Interface (WUI)
which has higher occruances of wildlifre events. Parameters include both weather variables, fire characteristics,
and characteristics of a home.
}
\examples{
home_risk_index(data_frame)
}
\author{
Hannah Garcia-Wickstrum and Anne-Marie Parkinson
}
