% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_min_temperature.R
\name{max_min_temp}
\alias{max_min_temp}
\title{Maximum and Minimum Temperatures}
\usage{
max_min_temp(df)
}
\arguments{
\item{df}{refers to the data frame which has a list of temperatures}
}
\value{
maximum and minimum temperatures
}
\description{
This function returns the maximum and minimum temperatures from a list
}
\examples{
max_min_temp(ex_df)
}
\author{
Anne-Marie and Hannah
}
