% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_winter_precip (1).R
\name{meanwinterprecip}
\alias{meanwinterprecip}
\title{Mean Winter Precipitation}
\usage{
meanwinterprecip(sum_winter_precip, no_of_samples)
}
\arguments{
\item{sum_winter_precip}{is the total winter precipitation (in inches)}

\item{no_of_samples}{is the number of locations where precipitation was recorded}
}
\value{
mean precipitation (inches)
}
\description{
This function computes the mean winter precipitation
}
\examples{
meanwinterprecip(225, 25)
}
\author{
Anne-Marie and Hannah
}
